add logic to play 1 full round of rock paper scissors
i added the logic to play a full round, while doing so i realized
my current logic for grabbing the computer input was making things
more difficult so i changed that as well. previously it returned
a number which represented rock paper or scissors but now i convert
that number into a string first so it returns a string same as the
human input.